@@ -1,4 +1,6 @@
-class Store < ActiveRecord::Base
-  attr_accessible :URL, :address, :email, :hoursOfOperation, :name, :phone
-  has_many :comments
+class Store < ApplicationRecord
+  has_many :comments, dependent: :destroy
+  
+  validates :name, presence: true
+  validates :email, format: { with: URI::MailTo::EMAIL_REGEXP }, allow_blank: true
 end
