@@ -1,6 +1,11 @@
-require 'rubygems'
+ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../Gemfile', __dir__)
 
-# Set up gems listed in the Gemfile.
-ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../../Gemfile', __FILE__)
+require 'bundler/setup' # Set up gems listed in the Gemfile.
 
-require 'bundler/setup' if File.exists?(ENV['BUNDLE_GEMFILE'])
+# Fix for mutex_m gem missing in Ruby 3.4.4
+require 'mutex_m'
+
+# Fix for logger gem conflict in Rails 6
+require 'logger'
+
+require 'bootsnap/setup' # Speed up boot time by caching expensive operations.
